+++ ImageOps/Blenders/BurnBlend.cs
@@ -2,9 +2,9 @@
 {
     public class BurnBlend : StandardBlend
     {
-        protected override float Blend(float color1, float color2)
+        protected override double Blend(double color1, double color2)
         {
-            return Clamp(1.0f - ((1.0f - color1)/color2));
+            return Clamp(1.0 - ((1.0 - color1)/color2));
         }
     }
 }
