+++ ImageOps/Blenders/GrainMergeBlend.cs
@@ -2,9 +2,9 @@
 {
     public class GrainMergeBlend : StandardBlend
     {
-        protected override float Blend(float color1, float color2)
+        protected override double Blend(double color1, double color2)
         {
-            return color1 + color2 - 0.5f;
+            return color1 + color2 - 0.5;
         }
     }
 }
