+++ ImageOps/Blenders/MultiplyBlend.cs
@@ -5,7 +5,7 @@
     /// </summary>
     public class MultiplyBlend : StandardBlend
     {
-        protected override float Blend(float color1, float color2)
+        protected override double Blend(double color1, double color2)
         {
             return color1*color2;
         }
